+++ flutter_quill_extensions/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## 0.6.5
 - Support the new improved platform checking of `flutter_quill`
+- Update the Image embed builder logic
+- Fix Save image button exception
+- Feature: Image cropping for the image embed builder
+- Add support for copying the image to the cliboard
+- Fix the image size logic (it's still missing a lot of things but we will work on that soon)
+- Fix the zoom image functionality to support different image providers
+- Update `README.md`
+- Deprecated: The boolean property `forceUseMobileOptionMenuForImageClick` is now deprecated as we will not using it anymore and it will be removed in the next major release
 
 ## 0.6.4
 - Update `QuillImageUtilities`